delete: unnecessary worker file
